requirements.txt
# dbt-core==1.7.0
# dbt-clickhouse==1.7.6
# python-decouple==3.8
# apache-airflow==2.10.5
# duckdb==1.2.2
# clickhouse-connect==0.7.19

requests==2.31.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlalchemy==1.4.52